support baning channels
Previously message from a channel in a group would be received from
a service bot with ID 136817688, and to ban a channel, you need
to pass its username to the ban command, which was not done.

This change starts using BanChatSenderChat for channels
instead of RestrictChatMember, which doesn't work on them.